server/requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
apache-airflow==2.7.1
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
APScheduler==3.10.4
argcomplete==3.1.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ckiptagger==0.2.1
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
community==1.0.0b1
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==15.1.0
contourpy==1.1.1
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
cycler==0.11.0
daphne==4.0.0
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
fake-useragent==1.2.1
fastapi==0.73.0
filelock==3.12.4
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
gast==0.4.0
gensim==4.3.2
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.2
hyperlink==21.0.0
idna==3.4
igraph==0.10.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
inflection==0.5.1
interchange==2021.0.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
keybert==0.7.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
libclang==16.0.6
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
loguru==0.7.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pansi==2020.7.3
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.0.1
pipdeptree==2.13.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.3.0
prison==0.2.1
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.7
py2neo==2021.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rich==13.5.3
rich-argparse==1.3.0
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
setproctitle==1.3.2
six==1.16.0
slowapi==0.1.8
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.17.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.2.3
wordcloud==1.9.2
wrapt==1.15.0
WTForms==3.0.1
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
